Init. Commit: Lake Pend Oreille Simple Stats
Small little program that, given a year in the range 2012-2016 (as of
this commit)
will retrieve, from https://lpo.dt.navy.mil/data/ , raw data and display
the mean and median of the wind speed. air temperature. and barometric
pressure.

Future Update:
--Will update to work with sqlite.
--As of now it parses the entire year's data but will be updated to work
on a monthly scale or with a date range.